Navigation close to working
