+++ toggles-app/src/main/AndroidManifest.xml
@@ -30,7 +30,7 @@
         tools:ignore="GoogleAppIndexingWarning">
 
         <activity
-            android:name=".BubbleActivity"
+            android:name=".bubble.BubbleActivity"
             android:theme="@style/AppTheme"
             android:label="@string/title_activity_bubble"
             android:allowEmbedded="true"
