+++ toggles-app/src/main/java/com/izettle/wrench/oss/LicenceMetadata.kt
@@ -1,7 +1,7 @@
 package com.izettle.wrench.oss
 
 import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
+import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class LicenceMetadata(
